fix: Update to version 2.31
