--- InventoryMaid/saves/startup.json
@@ -1,1 +1,0 @@
-0
